[jk] Clean up code
