+++ mage_ai/api/resources/PipelineRunResource.py
@@ -1,12 +1,12 @@
 from datetime import datetime
 from mage_ai.api.operations.constants import META_KEY_LIMIT, META_KEY_OFFSET
 from mage_ai.api.resources.DatabaseResource import DatabaseResource
-from mage_ai.data_integrations.utils.scheduler import initialize_state_and_runs
+from mage_ai.data_integrations.utils.scheduler import start_scheduler
 from mage_ai.data_preparation.models.constants import PipelineType
 from mage_ai.data_preparation.models.pipeline import Pipeline
 from mage_ai.orchestration.db import safe_db_query
 from mage_ai.orchestration.db.models.schedules import BlockRun, PipelineRun
-from mage_ai.orchestration.pipeline_scheduler import get_variables, stop_pipeline_run
+from mage_ai.orchestration.pipeline_scheduler import stop_pipeline_run
 from sqlalchemy.orm import selectinload
 
 
@@ -162,21 +162,8 @@ class PipelineRunResource(DatabaseResource):
             payload['create_block_runs'] = False
 
         def _create_callback(resource):
-            from mage_ai.orchestration.pipeline_scheduler import PipelineScheduler
-
             pipeline_run = resource.model
-            pipeline_scheduler = PipelineScheduler(pipeline_run)
-
-            if is_integration:
-                initialize_state_and_runs(
-                    pipeline_run,
-                    pipeline_scheduler.logger,
-                    get_variables(pipeline_run),
-                )
-            else:
-                pipeline_run.create_block_runs()
-
-            pipeline_scheduler.start(should_schedule=False)
+            start_scheduler(pipeline_run)
 
         self.on_create_callback = _create_callback
