+++ mage_ai/data_integrations/utils/scheduler.py
@@ -1,5 +1,6 @@
 from mage_ai.data_integrations.sources.constants import SQL_SOURCES
 from mage_ai.data_preparation.logging.logger import DictLogger
+from mage_ai.data_preparation.models.constants import PipelineType
 from mage_ai.data_preparation.models.pipelines.integration_pipeline import IntegrationPipeline
 from mage_ai.orchestration.db import db_connection
 from mage_ai.orchestration.db.models.schedules import BlockRun, PipelineRun
@@ -77,6 +78,24 @@ def initialize_state_and_runs(
         )
 
         raise err
+
+
+def start_scheduler(pipeline_run: PipelineRun) -> None:
+    from mage_ai.orchestration.pipeline_scheduler import PipelineScheduler, get_variables
+
+    pipeline_scheduler = PipelineScheduler(pipeline_run)
+    is_integration = PipelineType.INTEGRATION == pipeline_run.pipeline.type
+
+    if is_integration:
+        initialize_state_and_runs(
+            pipeline_run,
+            pipeline_scheduler.logger,
+            get_variables(pipeline_run),
+        )
+    else:
+        pipeline_run.create_block_runs()
+
+    pipeline_scheduler.start(should_schedule=False)
 
 
 def create_block_runs(pipeline_run: PipelineRun, logger: DictLogger) -> List[BlockRun]:
